requirements.txt
boto3==1.16.9
botocore==1.19.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
line-bot-sdk==1.17.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.21.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
